+++ TDEngine2/source/editor/IEditorWindow.cpp
@@ -41,6 +41,7 @@ namespace TDEngine2
 
 	void CBaseEditorWindow::_onUpdate(F32 dt)
 	{
+		mCurrDelta = dt;
 	}
 }